# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VAKSINA HR - Kadrlarni boshqaruv dasturi
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Barcha foydalanuvchilar
      parameters:
        - { name: role, in: query, schema: { type: string } }
        - { name: departmentId, in: query, schema: { type: number } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        "200":
          description: Foydalanuvchilar
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Yaratildi (temporaryPassword — bir marta ko'rsatiladi)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserResult"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      responses:
        "200":
          description: Foydalanuvchi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Yangilandi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      responses:
        "204":
          description: O'chirildi

  # --- REQUESTS (Zayavkalar) ---
...
components:
  schemas:
...
    User:
      type: object
      required: [id, fullName, role, status, createdAt]
      properties:
        id: { type: number }
        fullName: { type: string }
        role:
          type: string
          enum: [admin, recruiter, hr, hr_direktor, hr_auditor, hr_menejer, trainer, mentor, director, department_head, mudir, koordinator, texnik, ombor, farmasevt]
        departmentId: { type: ["integer", "null"] }
        departmentName: { type: ["string", "null"] }
        login: { type: string }
        phone: { type: ["string", "null"] }
        status:
          type: string
          enum: [active, inactive]
        createdAt: { type: string }
    UserInput:
      type: object
      required: [fullName, role]
      properties:
        fullName: { type: string }
        role: { type: string }
        departmentId: { type: ["integer", "null"] }
        login: { type: string }
        password: { type: string }
        phone: { type: ["string", "null"] }
        status: { type: string }
    CreateUserResult:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            temporaryPassword: { type: string }
    UserUpdate:
      type: object
      properties:
        fullName: { type: string }
        role: { type: string }
        departmentId: { type: ["integer", "null"] }
        phone: { type: ["string", "null"] }
        status: { type: string }

    # --- DEPARTMENT ---
...